requirements.txt
torch>=1.10.0
torchvision>=0.11.0
tqdm>=4.60.0
kagglehub>=0.1.0
Pillow>=8.0.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
streamlit>=1.0.0
torchsummary